EC-156 - OPENMRS-EIP-DOCKER: Implement download retry mechanism with integrity check

+++ release_stuff/etc/eptssync/scripts/eptssync_inconsistence_check.sh
@@ -11,9 +11,39 @@
 timestamp=`date +%Y-%m-%d_%H-%M-%S`
 
 EPTSSYNC_HOME="/home/eptssync"
+EIP_HOME="/home/eip"
 CONFIG_FILE=$EPTSSYNC_HOME/conf/source_sync_config.tmp.json
 CONFIG_FILE_ORIGINAL=$EPTSSYNC_HOME/conf/source_sync_config.json
-EIP_SCRIPTS_DIR="/home/eip_scripts"
+
+EIP_SETUP_SCRIPTS_DIR="$EIP_HOME/scripts"
+SHARED_DIR="$EIP_HOME/shared"
+RELEASES_PACKAGES_DIR="$SHARED_DIR/releases"
+APK_CMD=$(which apk)
+
+# Installing dependencies on APK distros and download release files if necessary
+if [ ! -z $APK_CMD ]
+then
+    apk add wget
+fi
+
+. $EIP_SETUP_SCRIPTS_DIR/release_info.sh
+echo "Verifying $RELEASE_NAME packages download status"
+$EIP_SETUP_SCRIPTS_DIR/download_release.sh "$RELEASES_PACKAGES_DIR" "$RELEASE_NAME" "$OPENMRS_EIP_APP_RELEASE_URL" "$EPTSSYNC_API_RELEASE_URL"
+
+CURRENT_RELEASES_PACKAGES_DIR="$RELEASES_PACKAGES_DIR/$RELEASE_NAME"
+
+RELEASE_PACKAGES_DOWNLOAD_COMPLETED="$CURRENT_RELEASES_PACKAGES_DIR/download_completed"
+if [ ! -f "$RELEASE_PACKAGES_DOWNLOAD_COMPLETED" ]
+then
+    echo "Error trying to download release packages: $RELEASE_NAME. See previous messages."
+    echo "EPTS startup process failed"
+    exit 1
+fi
+
+# copying files
+EPTSSYNC_PACKAGE_RELEASE_FILE_NAME=$(echo "$EPTSSYNC_API_RELEASE_URL" | rev | cut -d'/' -f 1 | rev)
+echo "Copying $EPTSSYNC_PACKAGE_RELEASE_FILE_NAME to $EPTSSYNC_HOME/eptssync-api-1.0-SNAPSHOT.jar"
+cp "$CURRENT_RELEASES_PACKAGES_DIR/$EPTSSYNC_PACKAGE_RELEASE_FILE_NAME" "$EPTSSYNC_HOME/eptssync-api-1.0-SNAPSHOT.jar"
 
 cp $CONFIG_FILE_ORIGINAL $CONFIG_FILE
 
@@ -33,9 +63,8 @@ then
    wget -O "$EPTSSYNC_HOME/$JAR_FILE_NAME" $EPTSSYNC_API_RELEASE_URL
 fi
 
-
 # Start application.
-echo -n "Starting EPTS Application"
+echo "Starting EPTS Application"
 cd $EPTSSYNC_HOME
 #java -jar eptssync-api-1.0-SNAPSHOT.jar "$CONFIG_FILE" 
 java -jar "$JAR_FILE_NAME" "$CONFIG_FILE" > $EPTSSYNC_HOME/logs/log.txt
